# [FEATURE] Adds performance test for affine alignment.