--- a/test/performance/alignment/CMakeLists.txt
+++ b/test/performance/alignment/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_benchmark(affine_alignment_benchmark.cpp)