Adds the first benchmark test for the alignment module:

Preliminary results for score only:

```
--------------------------------------------------------------
Benchmark                       Time           CPU Iterations
--------------------------------------------------------------
seqan3_affine_dna4         613038 ns     612646 ns       1015
seqan2_affine_dna4         590937 ns     590440 ns       1169
seqan3_affine_dna4_set   59400658 ns   59376083 ns         12
seqan2_affine_dna4_set   59811930 ns   59782583 ns         12
```